slot passback update mapping as actions
